+++ src/test/java/org/baeldung/test/ChangePasswordIntegrationTest.java
@@ -11,7 +11,7 @@ import java.util.Map;
 
 import org.baeldung.persistence.dao.UserRepository;
 import org.baeldung.persistence.model.User;
-import org.baeldung.Application;
+import org.baeldung.spring.Application;
 import org.baeldung.spring.TestDbConfig;
 import org.baeldung.spring.TestIntegrationConfig;
 import org.hamcrest.core.IsNot;
