+++ src/test/java/org/baeldung/test/GetLoggedUsersIntegrationTest.java
@@ -8,7 +8,7 @@ import java.util.Map;
 
 import org.baeldung.persistence.dao.UserRepository;
 import org.baeldung.persistence.model.User;
-import org.baeldung.Application;
+import org.baeldung.spring.Application;
 import org.baeldung.spring.TestDbConfig;
 import org.baeldung.spring.TestIntegrationConfig;
 import org.junit.Before;
